fix(serializers): rename avaliacao_serialiazer -> avaliacao_serializer em 4 arquivos
